+++ Sprint 3/ClientServerDemo/ClientGUI/Json Structure/Serializables/Sub Objects/Transform.cs
@@ -2,7 +2,7 @@
 {
     public class Transform
     {
-        public int[] position { get; set; }
+        public double[] position { get; set; }
         public double scale { get; set; }
         public int[] rotation { get; set; }
     }
